refactor: extract core vis func to package
Also rename the visualizer package to
visualizer-browser.
