+++ packages/visualizer-browser/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "visualizer",
+  "name": "@moonrepo/visualizer-browser",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "dependencies": {
-    "cytoscape": "3.23.0",
+    "@moonrepo/visualizer-core": "workspace:*",
     "preact": "10.11.2"
   },
   "devDependencies": {
     "@preact/preset-vite": "2.4.0",
-    "@types/cytoscape": "3.19.9",
     "eslint-config-preact": "1.3.0",
     "typescript": "4.6.4",
     "vite": "3.2.3",
