+++ packages/visualizer-core/README.md
@@ -1,0 +1,1 @@
+# visualizer-core
