+++ packages/visualizer-core/moon.yml
@@ -1,0 +1,2 @@
+language: typescript
+type: library
