+++ packages/visualizer-core/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.options.json",
+  "include": [
+    "**/*"
+  ],
+  "references": [],
+  "compilerOptions": {
+    "outDir": "../../.moon/cache/types/packages/visualizer-core"
+  }
+}
